# frontend/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Receituário API",
    "description" : "Modernized Receituário (migrated from GeneXus). Bearer JWT auth.",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:36219",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearer-jwt" : [ ]
  } ],
  "paths" : {
    "/api/usuarios/{usuCod}/unidades/{uniCod}" : {
      "get" : {
        "tags" : [ "Usuário — Unidades (permissões)" ],
        "summary" : "Obter a matriz de permissões de um usuário numa unidade",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "usuCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "uniCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UsuarioUnidadeResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Usuário — Unidades (permissões)" ],
        "summary" : "Atualizar as permissões de um usuário numa unidade",
        "operationId" : "update",
        "parameters" : [ {
          "name" : "usuCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "uniCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UsuarioUnidadeUpsertRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UsuarioUnidadeResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Usuário — Unidades (permissões)" ],
        "summary" : "Revogar o acesso de um usuário a uma unidade",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "usuCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "uniCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/usuarios/{id}" : {
      "get" : {
        "tags" : [ "Usuários (administração)" ],
        "summary" : "Obter usuário por código",
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Usuário",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Não encontrado",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UsuarioResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Usuários (administração)" ],
        "summary" : "Atualizar usuário (bloquear/desbloquear via bloqueado)",
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UsuarioUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UsuarioResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Usuários (administração)" ],
        "summary" : "Excluir usuário (bloqueado se referenciado por permissões/unidades)",
        "operationId" : "delete_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "409" : {
            "description" : "Usuário possui vínculos"
          },
          "204" : {
            "description" : "Excluído"
          }
        }
      }
    },
...
    "/api/unidades/{codigo}" : {
      "get" : {
        "tags" : [ "unidade-controller" ],
        "operationId" : "get_3",
        "parameters" : [ {
          "name" : "codigo",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UnidadeResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "unidade-controller" ],
        "operationId" : "update_3",
        "parameters" : [ {
          "name" : "codigo",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UnidadeUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UnidadeResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "unidade-controller" ],
        "operationId" : "delete_3",
        "parameters" : [ {
          "name" : "codigo",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/medicamentos/{remCod}/unidades/{uniCod}" : {
      "put" : {
        "tags" : [ "medicamento-sub-controller" ],
        "operationId" : "updateUnidade",
        "parameters" : [ {
          "name" : "remCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "uniCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UnidadeAtualizarRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UnidadeResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "medicamento-sub-controller" ],
        "operationId" : "removeUnidade",
        "parameters" : [ {
          "name" : "remCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "uniCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/usuarios/{usuCod}/unidades" : {
      "get" : {
        "tags" : [ "Usuário — Unidades (permissões)" ],
        "summary" : "Listar as unidades (e permissões) de um usuário",
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "usuCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UsuarioUnidadeResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Usuário — Unidades (permissões)" ],
        "summary" : "Conceder acesso de um usuário a uma unidade",
        "operationId" : "create_1",
        "parameters" : [ {
          "name" : "usuCod",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "uniCod",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UsuarioUnidadeUpsertRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UsuarioUnidadeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UsuarioUnidadeUpsertRequest" : {
        "type" : "object",
        "properties" : {
          "especialidadeCod" : {
            "type" : "integer",
            "format" : "int32"
          },
          "bloqueioTabela" : {
            "type" : "boolean"
          },
          "bloqueioCadastro" : {
            "type" : "boolean"
          },
          "bloqueioAmbulatorio" : {
            "type" : "boolean"
          },
          "permiteConsultaDireta" : {
            "type" : "boolean"
          },
          "bloqueioProntuarioConsulta" : {
            "type" : "boolean"
          },
          "bloqueioProntuarioOdonto" : {
            "type" : "boolean"
          },
          "bloqueioResultadoExame" : {
            "type" : "boolean"
          },
          "bloqueioEsus" : {
            "type" : "boolean"
          },
          "bloqueioCaps" : {
            "type" : "boolean"
          },
          "bloqueioNutricao" : {
            "type" : "boolean"
          },
          "bloqueioFarmacia" : {
            "type" : "boolean"
          },
          "permiteBnafar" : {
            "type" : "boolean"
          },
          "bloqueioAlmoxarifado" : {
            "type" : "boolean"
          },
          "bloqueioRequisicao" : {
            "type" : "boolean"
          },
          "bloqueioBeneficio" : {
            "type" : "boolean"
          },
          "bloqueioTransporte" : {
            "type" : "boolean"
          },
          "bloqueioVacina" : {
            "type" : "boolean"
          },
          "bloqueioAgenda" : {
            "type" : "boolean"
          },
          "bloqueioAgendaManual" : {
            "type" : "boolean"
          },
          "bloqueioAgendaExterna" : {
            "type" : "boolean"
          },
          "bloqueioAgendaEspecial" : {
            "type" : "boolean"
          },
          "permiteAgendaAuditor" : {
            "type" : "boolean"
          },
          "bloqueioLaboratorio" : {
            "type" : "boolean"
          },
          "bloqueioHospital" : {
            "type" : "boolean"
          },
          "bloqueioVigilancia" : {
            "type" : "boolean"
          },
          "bloqueioAgravo" : {
            "type" : "boolean"
          },
          "bloqueioCms" : {
            "type" : "boolean"
          },
          "bloqueioOuvidoria" : {
            "type" : "boolean"
          },
          "bloqueioImpressao" : {
            "type" : "boolean"
          },
          "bloqueioExportacao" : {
            "type" : "boolean"
          },
          "bloqueioParametro" : {
            "type" : "boolean"
          },
          "bloqueioRelatorio" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioTabela" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioCadastro" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioAmbulatorio" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioEsus" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioCaps" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioNutricao" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioVacina" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioFarmacia" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioAlmoxarifado" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioRequisicao" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioBeneficio" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioTransporte" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioAgenda" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioLaboratorio" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioHospital" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioVigilancia" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioAgravo" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioOuvidoria" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioExportacao" : {
            "type" : "boolean"
          },
          "bloqueioGrafico" : {
            "type" : "boolean"
          },
          "permiteAgendaAuditorPcd" : {
            "type" : "boolean"
          },
          "permiteSoaBnafar" : {
            "type" : "boolean"
          }
        }
      },
      "UsuarioUnidadeResponse" : {
        "type" : "object",
        "properties" : {
          "usuCod" : {
            "type" : "integer",
            "format" : "int32"
          },
          "uniCod" : {
            "type" : "integer",
            "format" : "int32"
          },
          "especialidadeCod" : {
            "type" : "integer",
            "format" : "int32"
          },
          "bloqueioTabela" : {
            "type" : "boolean"
          },
          "bloqueioCadastro" : {
            "type" : "boolean"
          },
          "bloqueioAmbulatorio" : {
            "type" : "boolean"
          },
          "permiteConsultaDireta" : {
            "type" : "boolean"
          },
          "bloqueioProntuarioConsulta" : {
            "type" : "boolean"
          },
          "bloqueioProntuarioOdonto" : {
            "type" : "boolean"
          },
          "bloqueioResultadoExame" : {
            "type" : "boolean"
          },
          "bloqueioEsus" : {
            "type" : "boolean"
          },
          "bloqueioCaps" : {
            "type" : "boolean"
          },
          "bloqueioNutricao" : {
            "type" : "boolean"
          },
          "bloqueioFarmacia" : {
            "type" : "boolean"
          },
          "permiteBnafar" : {
            "type" : "boolean"
          },
          "bloqueioAlmoxarifado" : {
            "type" : "boolean"
          },
          "bloqueioRequisicao" : {
            "type" : "boolean"
          },
          "bloqueioBeneficio" : {
            "type" : "boolean"
          },
          "bloqueioTransporte" : {
            "type" : "boolean"
          },
          "bloqueioVacina" : {
            "type" : "boolean"
          },
          "bloqueioAgenda" : {
            "type" : "boolean"
          },
          "bloqueioAgendaManual" : {
            "type" : "boolean"
          },
          "bloqueioAgendaExterna" : {
            "type" : "boolean"
          },
          "bloqueioAgendaEspecial" : {
            "type" : "boolean"
          },
          "permiteAgendaAuditor" : {
            "type" : "boolean"
          },
          "bloqueioLaboratorio" : {
            "type" : "boolean"
          },
          "bloqueioHospital" : {
            "type" : "boolean"
          },
          "bloqueioVigilancia" : {
            "type" : "boolean"
          },
          "bloqueioAgravo" : {
            "type" : "boolean"
          },
          "bloqueioCms" : {
            "type" : "boolean"
          },
          "bloqueioOuvidoria" : {
            "type" : "boolean"
          },
          "bloqueioImpressao" : {
            "type" : "boolean"
          },
          "bloqueioExportacao" : {
            "type" : "boolean"
          },
          "bloqueioParametro" : {
            "type" : "boolean"
          },
          "bloqueioRelatorio" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioTabela" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioCadastro" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioAmbulatorio" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioEsus" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioCaps" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioNutricao" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioVacina" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioFarmacia" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioAlmoxarifado" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioRequisicao" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioBeneficio" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioTransporte" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioAgenda" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioLaboratorio" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioHospital" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioVigilancia" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioAgravo" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioOuvidoria" : {
            "type" : "boolean"
          },
          "bloqueioRelatorioExportacao" : {
            "type" : "boolean"
          },
          "bloqueioGrafico" : {
            "type" : "boolean"
          },
          "permiteAgendaAuditorPcd" : {
            "type" : "boolean"
          },
          "permiteSoaBnafar" : {
            "type" : "boolean"
          }
        }
      },
      "UsuarioUpdateRequest" : {
        "required" : [ "nome" ],
        "type" : "object",
        "properties" : {
          "nome" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "tipo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "perfilId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "profissionalId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "funcionarioId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "bloqueado" : {
            "type" : "boolean"
          },
          "superusuario" : {
            "type" : "boolean"
          }
        }
      },
      "UsuarioResponse" : {
        "type" : "object",
        "properties" : {
          "usuCod" : {
            "type" : "integer",
            "format" : "int32"
          },
          "login" : {
            "type" : "string"
          },
          "nome" : {
            "type" : "string"
          },
          "tipo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "bloqueado" : {
            "type" : "boolean"
          },
          "perfilId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "superusuario" : {
            "type" : "boolean"
          },
          "profissionalId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "funcionarioId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ultimoAcesso" : {
            "type" : "string",
            "format" : "date"
          },
          "dataRedefinicaoSenha" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
...
      "UnidadeResponse" : {
        "type" : "object",
        "properties" : {
          "codigo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nome" : {
            "type" : "string"
          },
          "razaoSocial" : {
            "type" : "string"
          },
          "cnpj" : {
            "type" : "string"
          },
          "cep" : {
            "type" : "string"
          },
          "endereco" : {
            "type" : "string"
          },
          "enderecoNumero" : {
            "type" : "string"
          },
          "enderecoComplemento" : {
            "type" : "string"
          },
          "bairro" : {
            "type" : "string"
          },
          "telefone" : {
            "type" : "string"
          },
          "fax" : {
            "type" : "string"
          },
          "licencaFuncionamento" : {
            "type" : "string"
          },
          "responsavel" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "cnes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "bpa" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sipni" : {
            "type" : "integer",
            "format" : "int32"
          },
          "orgaoEmissor" : {
            "type" : "string"
          },
          "esferaAdministrativa" : {
            "type" : "integer",
            "format" : "int32"
          },
          "psf" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sisPreNatal" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hiperdia" : {
            "type" : "integer",
            "format" : "int32"
          },
          "gestao" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sia" : {
            "type" : "string"
          },
          "sigla" : {
            "type" : "string"
          },
          "situacao" : {
            "type" : "integer",
            "format" : "int32"
          },
          "siaSus" : {
            "type" : "string"
          },
          "scnesId" : {
            "type" : "string"
          },
          "exportarEsus" : {
            "type" : "boolean"
          },
          "exportarBnafar" : {
            "type" : "boolean"
          },
          "cadastroCns" : {
            "type" : "boolean"
          },
          "cadastroEndereco" : {
            "type" : "boolean"
          },
          "atendimentoSemCns" : {
            "type" : "boolean"
          },
          "atendimentoSemEndereco" : {
            "type" : "boolean"
          },
          "encaminhamentoFisioterapia" : {
            "type" : "boolean"
          },
          "externo" : {
            "type" : "boolean"
          },
          "forPesCod" : {
            "type" : "integer",
            "format" : "int64"
          },
          "tipoUnidadeCodigo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "atencaoSecundaria" : {
            "type" : "boolean"
          },
          "bloqueioPacSemCadInd" : {
            "type" : "boolean"
          },
          "avisoVacinaAtrasada" : {
            "type" : "boolean"
          },
          "cadastroCpf" : {
            "type" : "boolean"
          },
          "painel" : {
            "type" : "boolean"
          },
          "recepcaoIntermedMpp" : {
            "type" : "boolean"
          },
          "recepcaoIntermedMppImp" : {
            "type" : "boolean"
          },
          "baixaRemSemCns" : {
            "type" : "boolean"
          },
          "bloqueioLancPcdAut" : {
            "type" : "boolean"
          },
          "bloqueioDispPacExt" : {
            "type" : "boolean"
          },
          "bloqueioAgendSolExaPacExt" : {
            "type" : "boolean"
          },
          "municipioCodigo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "respProfissionalCodigo" : {
            "type" : "integer",
            "format" : "int64"
          },
          "diretorCodigo" : {
            "type" : "integer",
            "format" : "int64"
          },
          "auditorCodigo" : {
            "type" : "integer",
            "format" : "int64"
          },
          "autorizadorCodigo" : {
            "type" : "integer",
            "format" : "int64"
          },
          "distritoCodigo" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "UnidadeAtualizarRequest" : {
        "type" : "object",
        "properties" : {
          "estoqueMinimo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "situacao" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}